Add bug reporting functionality
This commit introduces a "Send Bug" button to the settings screen.
When clicked, the button triggers the `sendBugReport()` method in the service and displays a toast message confirming the report has been sent.

A new drawable `rounded_button.xml` has been added for the button's background.
The layout of `settings_activity.xml` has been adjusted to accommodate the new button.